chore: split watcher tests
