+++ watcher/package.json
@@ -8,10 +8,10 @@
     "codegen": "node generateEsModulesFromJson.cjs",
     "start": "node --require dotenv/config src/index.js",
     "test": "npm run test-token && npm run test-native",
-    "test-token": "echo 'RUNNING TESTS ON TOKEN CHAIN' && CHAIN=TOKEN CHAIN_ID=65535 IEXEC_ADDRESS=0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca ETH_RPC_HOST=http://localhost:8545 ETH_WS_HOST=http://localhost:8546 MONGO_HOST=localhost REDIS_HOST=localhost NODE_OPTIONS=--experimental-vm-modules jest watcher-common --runInBand --forceExit --detectOpenHandles --coverage",
-    "test-native": "echo 'RUNNING TESTS ON NATIVE CHAIN' && CHAIN=NATIVE IS_NATIVE=true CHAIN_ID=65535 IEXEC_ADDRESS=0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca ETH_RPC_HOST=http://localhost:18545 ETH_WS_HOST=http://localhost:18546 MONGO_HOST=localhost REDIS_HOST=localhost NODE_OPTIONS=--experimental-vm-modules jest watcher-common --runInBand --forceExit --detectOpenHandles --coverage",
-    "ci-test-token": "echo 'RUNNING TESTS ON TOKEN CHAIN' && CHAIN=TOKEN CHAIN_ID=65535 IEXEC_ADDRESS=0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca ETH_RPC_HOST=http://chain-token:8545 ETH_WS_HOST=http://chain-token:8546 MONGO_HOST=mongo REDIS_HOST=redis NODE_OPTIONS=--experimental-vm-modules jest watcher-common --runInBand --forceExit --detectOpenHandles --coverage",
-    "ci-test-native": "echo 'RUNNING TESTS ON NATIVE CHAIN' && CHAIN=NATIVE IS_NATIVE=true CHAIN_ID=65535 IEXEC_ADDRESS=0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca ETH_RPC_HOST=http://chain-native:8545 ETH_WS_HOST=http://chain-native:8546 MONGO_HOST=mongo REDIS_HOST=redis NODE_OPTIONS=--experimental-vm-modules jest watcher-common --runInBand --forceExit --detectOpenHandles --coverage",
+    "test-token": "echo 'RUNNING TESTS ON TOKEN CHAIN' && CHAIN=TOKEN CHAIN_ID=65535 IEXEC_ADDRESS=0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca ETH_RPC_HOST=http://localhost:8545 ETH_WS_HOST=http://localhost:8546 MONGO_HOST=localhost REDIS_HOST=localhost NODE_OPTIONS=--experimental-vm-modules jest --runInBand --forceExit --detectOpenHandles --coverage",
+    "test-native": "echo 'RUNNING TESTS ON NATIVE CHAIN' && CHAIN=NATIVE IS_NATIVE=true CHAIN_ID=65535 IEXEC_ADDRESS=0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca ETH_RPC_HOST=http://localhost:18545 ETH_WS_HOST=http://localhost:18546 MONGO_HOST=localhost REDIS_HOST=localhost NODE_OPTIONS=--experimental-vm-modules jest --runInBand --forceExit --detectOpenHandles --coverage",
+    "ci-test-token": "echo 'RUNNING TESTS ON TOKEN CHAIN' && CHAIN=TOKEN CHAIN_ID=65535 IEXEC_ADDRESS=0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca ETH_RPC_HOST=http://chain-token:8545 ETH_WS_HOST=http://chain-token:8546 MONGO_HOST=mongo REDIS_HOST=redis LOG_LEVEL=off NODE_OPTIONS=--experimental-vm-modules jest --runInBand --forceExit --detectOpenHandles --coverage",
+    "ci-test-native": "echo 'RUNNING TESTS ON NATIVE CHAIN' && CHAIN=NATIVE IS_NATIVE=true CHAIN_ID=65535 IEXEC_ADDRESS=0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca ETH_RPC_HOST=http://chain-native:8545 ETH_WS_HOST=http://chain-native:8546 MONGO_HOST=mongo REDIS_HOST=redis LOG_LEVEL=off NODE_OPTIONS=--experimental-vm-modules jest --runInBand --forceExit --detectOpenHandles --coverage",
     "debug-test": "export DEBUG=iexec-watcher* && npm run test",
     "stop-test-stack": "docker compose -f test/docker-compose.yml down -v",
     "start-test-stack": "npm run stop-test-stack && docker compose -f test/docker-compose.yml up -d"
